Crate dkg_runtime_primitives

Re-exports

pub use proposal::*;

pub use crate::proposal::DKGPayloadKey;

Modules

crypto

handlers

offchain

proposal

traits

util

Utility functions for Ethereum.

utils

Structs

AccessListItem

Account

AggregatedMisbehaviourReports

AggregatedPublicKeys

AuthoritySet

Block

Bloom

Bloom hash type with 256 bytes (2048 bits) size.

BloomRef

Commitment

EIP658ReceiptData

EIP1559Transaction

EIP1559TransactionMessage

EIP2930Transaction

EIP2930TransactionMessage

FromStrRadixErr

The error type for parsing numbers from strings.

FrontierReceiptData

H32

H64

H128

Fixed-size uninterpreted hash type with 16 bytes (128 bits) size.

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H264

H512

Fixed-size uninterpreted hash type with 64 bytes (512 bits) size.

H520

Header

Ethereum header definition.

LegacyTransaction

LegacyTransactionMessage

Log

OffchainSignedProposals

PartialHeader

Partial header definition without ommers hash and transactions root.

TransactionRecoveryId

TransactionSignature

U64

Little-endian large integer type Unsigned 64-bit integer.

U128

Little-endian large integer type 128-bit unsigned integer.

U256

Little-endian large integer type 256-bit unsigned integer.

U512

Little-endian large integer type 512-bits unsigned integer.

UnsignedProposal

Enums

BloomInput

ConsensusLog

DKGReport

FromDecStrErr

Conversion from decimal string error

FromStrRadixErrKind

A list of error categories encountered when parsing numbers.

MisbehaviourType

ReceiptAny

ReceiptV2

ReceiptV3

TransactionAction

TransactionV1

TransactionV2

Constants

DKG_ENGINE_ID

GENESIS_AUTHORITY_SET_ID

Authority set id starts with zero at genesis

GOSSIP_MESSAGE_RESENDING_LIMIT

Gossip message resending limit for outbound messages

KEYGEN_TIMEOUT

The keygen timeout limit in blocks before we consider misbehaviours

KEY_TYPE

OFFLINE_TIMEOUT

The offline timeout limit in blocks before we consider misbehaviours

SIGN_TIMEOUT

The sign timeout limit in blocks before we consider misbehaviours

UNTRACK_INTERVAL

Traits

BigEndianHash

DKGApi

Functions

keccak_256

Utility fn to calculate keccak 256 has

Type Definitions

AccessList

Address

AuthorityIndex

AuthoritySetId

BlockAny

BlockV0

BlockV1

BlockV2

EIP1559ReceiptData

EIP2930ReceiptData

MmrRootHash

The type used to represent an MMR root hash.

Public

PublicKeyAndSignature

ReceiptV0

ReceiptV1

Secret

Signature

TransactionAny

TransactionV0