Trait dkg_runtime_primitives::DKGApi

pub trait DKGApi<Block: BlockT, AuthorityId, N>: Core<Block> where
    AuthorityId: Codec + PartialEq,
    N: Codec + PartialEq + AtLeast32BitUnsigned, {

    fn authority_set(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<AuthoritySet<AuthorityId>, ApiError> { ... }
    fn authority_set_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<AuthoritySet<AuthorityId>, ApiError> { ... }
    fn get_best_authorities(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Vec<(u16, AuthorityId)>, ApiError> { ... }
    fn get_best_authorities_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Vec<(u16, AuthorityId)>, ApiError> { ... }
    fn get_next_best_authorities(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Vec<(u16, AuthorityId)>, ApiError> { ... }
    fn get_next_best_authorities_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Vec<(u16, AuthorityId)>, ApiError> { ... }
    fn signature_threshold(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u16, ApiError> { ... }
    fn signature_threshold_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u16, ApiError> { ... }
    fn keygen_threshold(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u16, ApiError> { ... }
    fn keygen_threshold_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u16, ApiError> { ... }
    fn next_signature_threshold(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u16, ApiError> { ... }
    fn next_signature_threshold_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u16, ApiError> { ... }
    fn next_keygen_threshold(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u16, ApiError> { ... }
    fn next_keygen_threshold_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u16, ApiError> { ... }
    fn queued_authority_set(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<AuthoritySet<AuthorityId>, ApiError> { ... }
    fn queued_authority_set_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<AuthoritySet<AuthorityId>, ApiError> { ... }
    fn should_refresh(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        _block_number: N
    ) -> Result<bool, ApiError> { ... }
    fn should_refresh_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        _block_number: N
    ) -> Result<bool, ApiError> { ... }
    fn next_dkg_pub_key(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Option<(AuthoritySetId, Vec<u8>)>, ApiError> { ... }
    fn next_dkg_pub_key_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Option<(AuthoritySetId, Vec<u8>)>, ApiError> { ... }
    fn dkg_pub_key(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<(AuthoritySetId, Vec<u8>), ApiError> { ... }
    fn dkg_pub_key_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(AuthoritySetId, Vec<u8>), ApiError> { ... }
    fn get_unsigned_proposals(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Vec<UnsignedProposal>, ApiError> { ... }
    fn get_unsigned_proposals_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Vec<UnsignedProposal>, ApiError> { ... }
    fn get_max_extrinsic_delay(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        block_number: N
    ) -> Result<N, ApiError> { ... }
    fn get_max_extrinsic_delay_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        block_number: N
    ) -> Result<N, ApiError> { ... }
    fn get_authority_accounts(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<(Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>, Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>), ApiError> { ... }
    fn get_authority_accounts_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>, Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>), ApiError> { ... }
    fn get_reputations(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        authorities: Vec<AuthorityId>
    ) -> Result<Vec<(AuthorityId, u128)>, ApiError> { ... }
    fn get_reputations_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        authorities: Vec<AuthorityId>
    ) -> Result<Vec<(AuthorityId, u128)>, ApiError> { ... }
    fn get_keygen_jailed(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        set: Vec<AuthorityId>
    ) -> Result<Vec<AuthorityId>, ApiError> { ... }
    fn get_keygen_jailed_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        set: Vec<AuthorityId>
    ) -> Result<Vec<AuthorityId>, ApiError> { ... }
    fn get_signing_jailed(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        set: Vec<AuthorityId>
    ) -> Result<Vec<AuthorityId>, ApiError> { ... }
    fn get_signing_jailed_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        set: Vec<AuthorityId>
    ) -> Result<Vec<AuthorityId>, ApiError> { ... }
    fn next_pub_key_sig(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Option<Vec<u8>>, ApiError> { ... }
    fn next_pub_key_sig_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Option<Vec<u8>>, ApiError> { ... }
    fn refresh_nonce(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u32, ApiError> { ... }
    fn refresh_nonce_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u32, ApiError> { ... }

}

Provided Methods

fn authority_set(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<AuthoritySet<AuthorityId>, ApiError>

Return the current active authority set

fn authority_set_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<AuthoritySet<AuthorityId>, ApiError>

Return the current active authority set

fn get_best_authorities(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Vec<(u16, AuthorityId)>, ApiError>

Return the current best authority set chosen for keygen

fn get_best_authorities_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Vec<(u16, AuthorityId)>, ApiError>

Return the current best authority set chosen for keygen

fn get_next_best_authorities(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Vec<(u16, AuthorityId)>, ApiError>

Return the next best authority set chosen for the queued keygen

fn get_next_best_authorities_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Vec<(u16, AuthorityId)>, ApiError>

Return the next best authority set chosen for the queued keygen

fn signature_threshold(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u16, ApiError>

Return the current signature threshold for the DKG

fn signature_threshold_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u16, ApiError>

Return the current signature threshold for the DKG

fn keygen_threshold(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u16, ApiError>

Return the current keygen threshold for the DKG

fn keygen_threshold_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u16, ApiError>

Return the current keygen threshold for the DKG

fn next_signature_threshold(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u16, ApiError>

Return the next signature threshold for the DKG

fn next_signature_threshold_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u16, ApiError>

Return the next signature threshold for the DKG

fn next_keygen_threshold(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u16, ApiError>

Return the next keygen threshold for the DKG

fn next_keygen_threshold_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u16, ApiError>

Return the next keygen threshold for the DKG

fn queued_authority_set(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<AuthoritySet<AuthorityId>, ApiError>

Return the next authorities active authority set

fn queued_authority_set_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<AuthoritySet<AuthorityId>, ApiError>

Return the next authorities active authority set

fn should_refresh(
&self,
__runtime_api_at_param__: &BlockId<Block>,
_block_number: N
) -> Result<bool, ApiError>

Check if refresh process should start

fn should_refresh_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
_block_number: N
) -> Result<bool, ApiError>

Check if refresh process should start

fn next_dkg_pub_key(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Option<(AuthoritySetId, Vec<u8>)>, ApiError>

Fetch DKG public key for queued authorities

fn next_dkg_pub_key_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Option<(AuthoritySetId, Vec<u8>)>, ApiError>

Fetch DKG public key for queued authorities

fn dkg_pub_key(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<(AuthoritySetId, Vec<u8>), ApiError>

Fetch DKG public key for current authorities

fn dkg_pub_key_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<(AuthoritySetId, Vec<u8>), ApiError>

Fetch DKG public key for current authorities

fn get_unsigned_proposals(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Vec<UnsignedProposal>, ApiError>

Get list of unsigned proposals

fn get_unsigned_proposals_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Vec<UnsignedProposal>, ApiError>

Get list of unsigned proposals

fn get_max_extrinsic_delay(
&self,
__runtime_api_at_param__: &BlockId<Block>,
block_number: N
) -> Result<N, ApiError>

Get maximum delay before which an offchain extrinsic should be submitted

fn get_max_extrinsic_delay_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
block_number: N
) -> Result<N, ApiError>

Get maximum delay before which an offchain extrinsic should be submitted

fn get_authority_accounts(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<(Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>, Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>), ApiError>

Current and Queued Authority Account Ids [/current_authorities/, /next_authorities/]

fn get_authority_accounts_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<(Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>, Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>), ApiError>

Current and Queued Authority Account Ids [/current_authorities/, /next_authorities/]

fn get_reputations(
&self,
__runtime_api_at_param__: &BlockId<Block>,
authorities: Vec<AuthorityId>
) -> Result<Vec<(AuthorityId, u128)>, ApiError>

Reputations for authorities

fn get_reputations_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
authorities: Vec<AuthorityId>
) -> Result<Vec<(AuthorityId, u128)>, ApiError>

Reputations for authorities

fn get_keygen_jailed(
&self,
__runtime_api_at_param__: &BlockId<Block>,
set: Vec<AuthorityId>
) -> Result<Vec<AuthorityId>, ApiError>

Returns the set of jailed keygen authorities from a set of authorities

fn get_keygen_jailed_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
set: Vec<AuthorityId>
) -> Result<Vec<AuthorityId>, ApiError>

Returns the set of jailed keygen authorities from a set of authorities

fn get_signing_jailed(
&self,
__runtime_api_at_param__: &BlockId<Block>,
set: Vec<AuthorityId>
) -> Result<Vec<AuthorityId>, ApiError>

Returns the set of jailed signing authorities from a set of authorities

fn get_signing_jailed_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
set: Vec<AuthorityId>
) -> Result<Vec<AuthorityId>, ApiError>

Returns the set of jailed signing authorities from a set of authorities

fn next_pub_key_sig(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Option<Vec<u8>>, ApiError>

Fetch DKG public key for sig

fn next_pub_key_sig_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Option<Vec<u8>>, ApiError>

Fetch DKG public key for sig

fn refresh_nonce(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u32, ApiError>

Get next nonce value for refresh proposal

fn refresh_nonce_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u32, ApiError>

Get next nonce value for refresh proposal