Trait dkg_runtime_primitives::DKGApi

pub trait DKGApi<Block: BlockT, AuthorityId, N, MaxProposalLength, MaxAuthorities>: Core<Block>where
    AuthorityId: Codec + PartialEq,
    MaxProposalLength: Get<u32> + Clone,
    MaxAuthorities: Get<u32> + Clone,
    N: Codec + PartialEq + AtLeast32BitUnsigned,{

    // Provided methods
    fn authority_set(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<AuthoritySet<AuthorityId, MaxAuthorities>, ApiError> { ... }
    fn get_best_authorities(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Vec<(u16, AuthorityId)>, ApiError> { ... }
    fn get_next_best_authorities(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Vec<(u16, AuthorityId)>, ApiError> { ... }
    fn get_current_session_progress(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        block_number: N
    ) -> Result<Option<Permill>, ApiError> { ... }
    fn signature_threshold(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<u16, ApiError> { ... }
    fn keygen_threshold(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<u16, ApiError> { ... }
    fn next_signature_threshold(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<u16, ApiError> { ... }
    fn next_keygen_threshold(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<u16, ApiError> { ... }
    fn queued_authority_set(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<AuthoritySet<AuthorityId, MaxAuthorities>, ApiError> { ... }
    fn should_refresh(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        _block_number: N
    ) -> Result<bool, ApiError> { ... }
    fn next_dkg_pub_key(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Option<(AuthoritySetId, Vec<u8>)>, ApiError> { ... }
    fn dkg_pub_key(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<(AuthoritySetId, Vec<u8>), ApiError> { ... }
    fn get_unsigned_proposal_batches(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Vec<StoredUnsignedProposalBatch<BatchId, MaxProposalLength, MaxProposalsInBatch, N>>, ApiError> { ... }
    fn get_authority_accounts(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<(Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>, Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>), ApiError> { ... }
    fn get_reputations(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        authorities: Vec<AuthorityId>
    ) -> Result<Vec<(AuthorityId, u128)>, ApiError> { ... }
    fn get_keygen_jailed(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        set: Vec<AuthorityId>
    ) -> Result<Vec<AuthorityId>, ApiError> { ... }
    fn get_signing_jailed(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        set: Vec<AuthorityId>
    ) -> Result<Vec<AuthorityId>, ApiError> { ... }
    fn next_pub_key_sig(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Option<Vec<u8>>, ApiError> { ... }
    fn refresh_nonce(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<u32, ApiError> { ... }
    fn should_execute_new_keygen(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<(bool, bool), ApiError> { ... }
    fn should_submit_proposer_vote(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<bool, ApiError> { ... }

}

Provided Methods

fn authority_set( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<AuthoritySet<AuthorityId, MaxAuthorities>, ApiError>

Return the current active authority set

fn get_best_authorities( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Vec<(u16, AuthorityId)>, ApiError>

Return the current best authority set chosen for keygen

fn get_next_best_authorities( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Vec<(u16, AuthorityId)>, ApiError>

Return the next best authority set chosen for the queued keygen

fn get_current_session_progress( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, block_number: N ) -> Result<Option<Permill>, ApiError>

Returns the progress of current session

fn signature_threshold( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<u16, ApiError>

Return the current signature threshold for the DKG

fn keygen_threshold( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<u16, ApiError>

Return the current keygen threshold for the DKG

fn next_signature_threshold( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<u16, ApiError>

Return the next signature threshold for the DKG

fn next_keygen_threshold( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<u16, ApiError>

Return the next keygen threshold for the DKG

fn queued_authority_set( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<AuthoritySet<AuthorityId, MaxAuthorities>, ApiError>

Return the next authorities active authority set

fn should_refresh( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, _block_number: N ) -> Result<bool, ApiError>

Check if refresh process should start

fn next_dkg_pub_key( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Option<(AuthoritySetId, Vec<u8>)>, ApiError>

Fetch DKG public key for queued authorities

fn dkg_pub_key( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<(AuthoritySetId, Vec<u8>), ApiError>

Fetch DKG public key for current authorities

fn get_unsigned_proposal_batches( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Vec<StoredUnsignedProposalBatch<BatchId, MaxProposalLength, MaxProposalsInBatch, N>>, ApiError>

Get list of unsigned proposals

fn get_authority_accounts( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<(Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>, Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId>), ApiError>

Current and Queued Authority Account Ids [/current_authorities/, /next_authorities/]

fn get_reputations( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, authorities: Vec<AuthorityId> ) -> Result<Vec<(AuthorityId, u128)>, ApiError>

Reputations for authorities

fn get_keygen_jailed( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, set: Vec<AuthorityId> ) -> Result<Vec<AuthorityId>, ApiError>

Returns the set of jailed keygen authorities from a set of authorities

fn get_signing_jailed( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, set: Vec<AuthorityId> ) -> Result<Vec<AuthorityId>, ApiError>

Returns the set of jailed signing authorities from a set of authorities

fn next_pub_key_sig( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Option<Vec<u8>>, ApiError>

Fetch DKG public key for sig

fn refresh_nonce( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<u32, ApiError>

Get next nonce value for refresh proposal

fn should_execute_new_keygen( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<(bool, bool), ApiError>

Returns (true, false) if we should execute a new keygen. Returns (true, true) if we should execute a forced new keygen.

fn should_submit_proposer_vote( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<bool, ApiError>

Whether to submit proposer vote